requirements.txt
hyperopt==0.2.3
lightgbm==3.2.1
matplotlib==3.4.3
numpy==1.21.2
pandas===1.4.1
pangres==2.3.1
pyclustertend==1.4.9
python_dateutil==2.8.2
s_dbw==0.4.0
scikit_learn==1.0.1
scipy==1.8.0
seaborn==0.11.2
slack_sdk==3.11.2
SQLAlchemy==1.3.18
statsmodels==0.13.0
yagmail==0.14.260
psycopg2-binary==2.9.2
tqdm==4.62.3
retry==0.9.2
xlsxwriter==3.0.2
modin==0.14.0
coloredlogs==15.0.1
humanfriendly==10.0